api_keys: move APIKeyForm to Views/, collapse namespace declarations
Demonstrates the views/ vs components/ rule (#4324) on a component
that's only used by one controller. Components::APIKeyForm becomes
Views::Controllers::Account::APIKeys::Form. The three callers
(table.rb, new.html.erb, edit.html.erb) all live under the same
api_keys controller; with the move, table.rb can reference its
sibling as just `Form`.

Also collapses the 4-level deep `module Views; module Controllers;
module Account; module APIKeys` opens into a single-line
`module Views::Controllers::Account::APIKeys` declaration — same
Ruby semantics, much less indentation. Applied to form.rb, table.rb,
and both view test files.
