+++ app/views/controllers/account/api_keys/new.html.erb
@@ -8,7 +8,7 @@ add_context_nav(account_api_tabs)
 
 <%= :account_api_keys_help.tp %>
 
-<%= render(Components::APIKeyForm.new(
+<%= render(Views::Controllers::Account::APIKeys::Form.new(
              @key,
              action: account_api_keys_path,
              id: "new_api_key_form")) %>
